refactor(searchPage): split scss file
