+++ packages/mirror-tv-next/components/search/_styles/no-search-result.module.scss
@@ -18,79 +18,6 @@
   }
 }
 
-.searchResultList {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  align-items: center;
-  gap: 28px;
-
-  @include media-breakpoint-up(md) {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-
-    row-gap: 48px;
-    column-gap: 20px;
-  }
-
-  @include media-breakpoint-up(lg) {
-    grid-template-columns: repeat(4, 1fr);
-    row-gap: 40px;
-    column-gap: 24px;
-  }
-
-  @include media-breakpoint-up(xl) {
-    row-gap: 52px;
-    column-gap: 24px;
-  }
-
-  li {
-    width: 100%;
-  }
-}
-
-.searchKeyword {
-  width: 100%;
-  margin-bottom: 20px;
-  text-align: left;
-  font-weight: 500;
-  font-size: 18px;
-  color: #014db8;
-}
-
-.seeMoreBtn {
-  width: 100%;
-  height: 37px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 3px solid #014db8;
-  color: #014db8;
-  font-size: 20px;
-  font-weight: 500;
-  letter-spacing: 0.5px;
-
-  @include media-breakpoint-up(md) {
-    width: 400px;
-    justify-self: center;
-    border-width: 1px;
-    grid-column: span 3;
-    letter-spacing: 3px;
-  }
-
-  @include media-breakpoint-up(lg) {
-    grid-column: span 4;
-  }
-}
-
-.noResultImg {
-  @include media-breakpoint-up(md) {
-    width: 200px;
-    height: 246px;
-    margin-top: 49px;
-  }
-}
-
 .noResultTxt {
   display: flex;
   flex-direction: column;
@@ -108,6 +35,14 @@
   :first-child {
     color: #898f9c;
     margin-bottom: 3px;
+  }
+}
+
+.noResultImg {
+  @include media-breakpoint-up(md) {
+    width: 200px;
+    height: 246px;
+    margin-top: 49px;
   }
 }
 
@@ -177,8 +112,14 @@
     grid-template-columns: repeat(4, 1fr);
     gap: 24px;
   }
-  li {
+
+  a {
+    display: block;
     width: 100%;
     height: 100%;
+    li {
+      width: 100%;
+      height: 100%;
+    }
   }
 }
